Implement parsing of indexing statistics for xia2 with DIALS
The AutoED now searches for the indexing statistics (the number
of indexed and the total number of spots) in the DIALS log
files. It parses these two numbers, and adds them to the json
database. The Javascript report generator is updated to read
those from the database and populate the report summary table.